Make the whole footer content more customizable
Refs #327.
