+++ app/components/layout/__tests__/Footer.spec.js
@@ -1,15 +1,22 @@
 import { expect } from 'chai';
 import React from 'react';
-import sd from 'skin-deep';
+import { shallow } from 'enzyme';
 
 import Footer from 'components/layout/Footer';
+import FooterContent from 'components/customization/FooterContent';
 
 describe('Component: layout/Footer', () => {
-  const props = {};
-  const tree = sd.shallowRender(<Footer {...props} />);
+  function getWrapper() {
+    return shallow(<Footer />);
+  }
 
   it('should render a footer element', () => {
-    const footerTrees = tree.everySubTree('footer');
-    expect(footerTrees.length).to.equal(1);
+    const footer = getWrapper().find('footer');
+    expect(footer.length).to.equal(1);
+  });
+
+  it('should render FooterContent component', () => {
+    const footerContent = getWrapper().find(FooterContent);
+    expect(footerContent.length).to.equal(1);
   });
 });
